+++ library/src/main/java/com/pengrad/telegrambot/request/ReopenForumTopic.java
@@ -3,7 +3,7 @@ package com.pengrad.telegrambot.request;
 import com.pengrad.telegrambot.response.BaseResponse;
 
 public class ReopenForumTopic extends BaseRequest<ReopenForumTopic, BaseResponse> {
-    public ReopenForumTopic(Integer chatId, Integer messageThreadId) {
+    public ReopenForumTopic(Long chatId, Integer messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
