+++ library/src/main/java/com/pengrad/telegrambot/request/UnpinAllForumTopicMessages.java
@@ -3,7 +3,7 @@ package com.pengrad.telegrambot.request;
 import com.pengrad.telegrambot.response.BaseResponse;
 
 public class UnpinAllForumTopicMessages extends BaseRequest<UnpinAllForumTopicMessages, BaseResponse> {
-    public UnpinAllForumTopicMessages(Integer chatId, Integer messageThreadId) {
+    public UnpinAllForumTopicMessages(Long chatId, Integer messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
